requirements.txt
streamlit 
google-generativeai
selenium
webdriver-manager
beautifulsoup4
lxml 
html5lib
python-dotenv